fixed upgrade erros
