+++ src/components/AddDoctor/Adddoctor.tsx
@@ -2,12 +2,20 @@ import {Department, DepartmentObject, Doctor, WorkingTime} from "../../api/model
 import {AuthComponent, AuthPropsLoc, AuthState} from "../../api/auth";
 
 import {withRouter} from "react-router";
-import {MenuItem, TextField} from "@material-ui/core";
-import {Button} from "@material-ui/core";
-import {Avatar} from '@material-ui/core';
-import React from "react";
+import React, { Component } from "react";
 import close from "../../images/close.svg";
+
 import {Skeleton} from "antd";
+import LocalizationProvider from "@mui/lab/LocalizationProvider";
+import TableContainer from "@mui/material/TableContainer";
+import {Table, TableHead, TableRow, TableCell, TableBody, FormControl, Button } from "@mui/material";
+import Select from "@mui/material/Select";
+import MenuItem from "@mui/material/MenuItem";
+import MobileTimePicker from "@mui/lab/MobileTimePicker";
+import TextField from "@mui/material/TextField";
+import Avatar from "@mui/material/Avatar";
+import AdapterDateFns from '@mui/lab/AdapterDateFns';
+
 
 interface AddDoctorState extends AuthState {
     name: string,
@@ -20,6 +28,69 @@ interface AddDoctorState extends AuthState {
     allDepartments: Array<DepartmentObject>,
     hospital: Array<number>,
     ready?: boolean
+}
+
+export class ResponsiveTimePickers extends Component<{}, { value: number | null,items:number[] }> {
+
+    days = ["sunday", "monday", "Tuesday", "wednesday", "thursday", "friday", "saturday"]
+
+    render() {
+        return (
+            <LocalizationProvider dateAdapter={AdapterDateFns}>
+                <TableContainer>
+                    <Table sx={{minWidth: 200}}>
+                        <TableHead>
+                            <TableRow>
+                                <TableCell>day</TableCell>
+                                <TableCell>Starting time</TableCell>
+                                <TableCell>Ending time</TableCell>
+                            </TableRow>
+                        </TableHead>
+                        <TableBody>
+                            {this.state.items.map((item, key) => (
+
+                                <TableRow>
+                                    <TableCell component="th" scope="row">
+                                        <FormControl>
+
+                                            <Select
+                                                style={{width: "120px"}}
+                                                labelId="demo-simple-select-label"
+                                                id="demo-simple-select"
+                                            >
+                                                {this.days.map((day, key) => (
+                                                    <MenuItem value={key}>{day}</MenuItem>)
+                                                )
+                                                }
+
+                                            </Select>
+                                        </FormControl>
+                                    </TableCell>
+                                    <TableCell align="right"> <MobileTimePicker
+
+                                        value={this.state.value}
+                                        onChange={(value) => {
+                                            this.setState({value});
+                                        }}
+                                        renderInput={(params) => <TextField {...params} />}
+                                    /></TableCell>
+                                    <TableCell align="right"> <MobileTimePicker
+
+                                        value={this.state.value}
+                                        onChange={(value) => {
+                                            this.setState({value});
+                                        }}
+                                        renderInput={(params) => <TextField {...params} />}
+                                    /></TableCell>
+
+                                </TableRow>
+                            ))}
+                        </TableBody>
+                    </Table>
+                </TableContainer>
+            </LocalizationProvider>
+        );
+    }
 }
 
 class AddDoctor extends AuthComponent<AuthPropsLoc, AddDoctorState> {
